So many fucking things dude

+++ src/main/java/com/eazybytes/springsecsection1/repository/CustomerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<CustomerEntity, Long> {
+public interface CustomerRepository extends CrudRepository<CustomerEntity, Integer> {
     Optional<CustomerEntity> findByEmail(String email);
 
     Optional<CustomerEntity> findByEmailAndPwd(String email, String pwd);
